+++ src/memoryfm/io/api.py
@@ -3,5 +3,6 @@ Data IO api
 """
 
 from memoryfm.io.lastfmstats import from_lastfmstats
+from memoryfm.io.spotify import from_spotify
 
-__all__ = ["from_lastfmstats"]
+__all__ = ["from_lastfmstats", "from_spotify"]